+++ sentry-spring-boot-starter/src/test/kotlin/io/sentry/spring/boot/SentryAutoConfigurationTest.kt
@@ -95,8 +95,8 @@ class SentryAutoConfigurationTest {
             "sentry.release=1.0.3",
             "sentry.environment=production",
             "sentry.sample-rate=0.2",
-            "sentry.in-app-excludes[0]=org.springframework",
-            "sentry.in-app-includes[0]=com.myapp",
+            "sentry.in-app-includes=org.springframework,com.myapp",
+            "sentry.in-app-excludes=org.jboss,com.microsoft",
             "sentry.dist=my-dist",
             "sentry.attach-threads=true",
             "sentry.attach-stacktrace=true",
@@ -121,8 +121,8 @@ class SentryAutoConfigurationTest {
             assertThat(options.release).isEqualTo("1.0.3")
             assertThat(options.environment).isEqualTo("production")
             assertThat(options.sampleRate).isEqualTo(0.2)
-            assertThat(options.inAppExcludes).containsOnly("org.springframework")
-            assertThat(options.inAppIncludes).containsOnly("com.myapp")
+            assertThat(options.inAppIncludes).containsOnly("org.springframework", "com.myapp")
+            assertThat(options.inAppExcludes).containsOnly("com.microsoft", "org.jboss")
             assertThat(options.dist).isEqualTo("my-dist")
             assertThat(options.isAttachThreads).isEqualTo(true)
             assertThat(options.isAttachStacktrace).isEqualTo(true)