+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Enhancement: Add API validator and remove deprecated methods
 * Enhancement: Add more convenient method to start a child span (#1073)
 * Enhancement: Autoconfigure traces callback in Spring Boot integration (#1074)
+* Enhancement: Resolve in-app-includes and in-app-excludes parameters from the external configuration
 
 # 4.0.0-alpha.1